dd_exterior/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.15
django-allauth==0.61.1
django-ckeditor==6.7.1
django-debug-toolbar==4.3.0
django-js-asset==2.2.0
django-livereload-server==0.5.1
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
oauthlib==3.2.2
outcome==1.3.0.post0
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0